--- a/lib/globalize.js
+++ b/lib/globalize.js
@@ -136,14 +136,21 @@ function formatMessage(path, variables, lang) {
     debug(
       '*** %s for %s not localized.  Fall back to English. ***  %s',
       path, lang, e);
-    try {
-      message = formatMsgInline(helper.ENGLISH);
-    } catch (e) {
+    if (lang === helper.ENGLISH) {
       message = sanitizeMsg(message, variables);
       debug(
         '*** %s not found for %s.  Fall back to: "%s" ***  %s',
         path, lang, message, e);
-    };
+    } else {
+      try {
+        message = formatMsgInline(helper.ENGLISH);
+      } catch (e) {
+        message = sanitizeMsg(message, variables);
+        debug(
+          '*** %s not found for %s.  Fall back to: "%s" ***  %s',
+          path, lang, message, e);
+      };
+    }
   };
   if (global.STRONGLOOP_GLB.PSEUDO_LOC_PREAMBLE) {
     message = global.STRONGLOOP_GLB.PSEUDO_LOC_PREAMBLE + message;