--- a/lib/globalize.js
+++ b/lib/globalize.js
@@ -133,21 +133,21 @@ function formatMessage(path, variables, lang) {
   try {
     message = formatMsgInline(lang);
   } catch (e) {
-    debug(
-      '*** %s for %s not localized.  Fall back to English. ***  %s',
-      path, lang, e);
     if (lang === helper.ENGLISH) {
       message = sanitizeMsg(message, variables);
       debug(
-        '*** %s not found for %s.  Fall back to: "%s" ***  %s',
+        '*** %s not found for %s. Fall back to: "%s" ***  %s',
         path, lang, message, e);
     } else {
+      debug(
+        '*** %s for %s not localized. Fall back to English. ***  %s',
+      path, lang, e);
       try {
         message = formatMsgInline(helper.ENGLISH);
       } catch (e) {
         message = sanitizeMsg(message, variables);
         debug(
-          '*** %s not found for %s.  Fall back to: "%s" ***  %s',
+          '*** %s not found for %s. Fall back to: "%s" ***  %s',
           path, lang, message, e);
       };
     }
@@ -204,7 +204,8 @@ function packMessage(args, fn, withOriginalMsg, lang) {
   // but there can be one or more %s in the message value.
   var variables = percentInKey ? helper.mapArgs(path, args) :
     txtWithTwoOrMoreArgs ? helper.repackArgs(args, 1) : args[1];
-  debug('packMessage:', path, JSON.stringify(args, null, 2), variables);
+  if (debug.enabled)
+    debug('packMessage:', path, JSON.stringify(args, null, 2), variables);
   var message = formatMessage(path, variables, lang);
   if (withOriginalMsg) message = {
     language: lang,