--- a/lib/lint.js
+++ b/lib/lint.js
@@ -111,7 +111,8 @@ function lintMessageFiles(enOnly, callback) {
   }
 
   function checkTrait(trait) {
-    debug('checkTrait:', JSON.stringify(trait, null, 4));
+    if (debug.enabled)
+      debug('checkTrait:', JSON.stringify(trait, null, 4));
     var verified = true;
     var plural = false;
     if (trait.nFiles === 0) {