This is a smaller version of #65, I hope the changes are small enough to be good to merge.

The performance before this change:

```
Expected 1940 calls of "g.f()" to finish under 3040.8940000000002ms, they took 1456.988444ms instead.
```

The performance after this change:

```
Expected 1940 calls of "g.f()" to finish under 1461.3521999999998ms, they took 848.279503ms instead.
```

Note that the ratio in the unit-test has to be much higher, because the tests run with coverage enabled, which slows down the globalize algorithm even more.

@Setogit please review

Connect to #66