@bajtos 2016-08-12
The timing (performance) test was rather unreliable, I have reworked it into a benchmark that has to be run manually.

Baseline performance across 3 runs

```
g.f() is 524x slower than util.format
g.f() is 474x slower than util.format
g.f() is 559x slower than util.format
```

With the patch in place:

```
g.f() is 328x slower than util.format
g.f() is 298x slower than util.format
g.f() is 362x slower than util.format
```