@bajtos 2016-08-15
Released as `strong-globalize@2.6.7`